backend/system_service/pyproject.toml
[tool.poetry]
name = "transportik-lite"
version = "0.0.1"
description = ""
authors = ["Андрей Романов <romanov_andrey_vl@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
markdown = "^3.1"
django-filter = "^2.2"
django-split-settings = "^1.0"
psycopg2 = "^2.8"
gunicorn = "^19.9"
layered-yaml-attrdict-config = "^18.12"
djangorestframework-jwt = "^1.11"
bcrypt = "^3.1"
coreapi = "^2.3"
coreschema = "^0.0.4"
djangorestframework-gis = "^0.15"
cassandra-driver = "^3.21"
lz4 = "^3.0"
celery = "^4.4"
django-celery-beat = "^1.5"
pika = "^1.1"
redis = "^3.3.11"
pydruid = "^0.5.8"
pandas = "^0.25.3"
django-pandas = "^0.6.1"
numpy = "^1.18.1"
pyspark = "^2.4.4"
clickhouse-driver = "^0.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pytest-django = "^3.8"
pytest-cov = "^2.8"
pytest-randomly = "^3.2"
pytest-testmon = "^1.0"
pytest-timeout = "^1.3"
jupyter = "^1.0.0"
ipython = "^7.11.1"
django-extensions = "^2.2.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
